refactor(package): keep graph loading behind cargo boundary
