+++ crates/sifr_package/src/graph/mod.rs
@@ -6,9 +6,6 @@ mod digest_cargo_metadata;
 mod digest_package_graph;
 mod digest_source_map;
 pub mod filters;
-mod load;
 pub mod scopes;
 pub mod type_identity;
 pub mod workspace;
-
-pub use load::{load_package_graph_snapshot, PackageGraphSnapshot};
