+++ crates/sifr_package/src/lib.rs
@@ -15,6 +15,7 @@ pub use crate::cargo::commands::{
     CargoPackageSelection, CargoPublishOptions, CargoVendorOptions,
 };
 pub use crate::cargo::errors::{map_cargo_failure, CargoAction};
+pub use crate::cargo::load::{load_package_graph_snapshot, PackageGraphSnapshot};
 pub use crate::cargo::lock_modes::{validate_offline_source_availability, CargoLockMode};
 pub use crate::cargo::metadata::{
     parse_metadata_json, CargoDependency, CargoMetadata, CargoPackage, CargoPackageId,
@@ -48,7 +49,6 @@ pub use crate::graph::workspace::{
     explicit_package_selection, select_sifr_workspace_members, selected_workspace_members,
     WorkspacePackageSelection,
 };
-pub use crate::graph::{load_package_graph_snapshot, PackageGraphSnapshot};
 pub use crate::imports::source_map::{
     DottedModulePath, PackageImportAmbiguity, PackageImportOrigin, PackageImportResolution,
     PackageImportResolutionResult, PackageModuleKey, PackageModuleSource, PackageSourceMap,
